refactor(platforms): consolidate Feishu platform implementation
- Implements #194 - Phase 2 of the unified file transfer system refactoring
- Moves Feishu platform components from channels/platforms/feishu/ to platforms/feishu/
- Updates all import paths to use the new location
- Maintains backward compatibility through @deprecated re-exports in channels/platforms/feishu/index.ts

## Changes

### Moved Files
- channels/platforms/feishu/feishu-adapter.ts → platforms/feishu/feishu-adapter.ts
- channels/platforms/feishu/feishu-message-sender.ts → platforms/feishu/feishu-message-sender.ts
- channels/platforms/feishu/feishu-file-handler.ts → platforms/feishu/feishu-file-handler.ts
- channels/platforms/feishu/card-builders/ → platforms/feishu/card-builders/
- Test files moved accordingly

### Updated Imports
- Updated relative import paths in moved files
- Updated imports in feishu-channel.ts to use new location

## Test Results

Test Files  43 passed (43)
Tests       716 passed (716)
