+++ src/channels/feishu-channel.ts
@@ -11,8 +11,8 @@ import { DEDUPLICATION, REACTIONS } from '../config/constants.js';
 import { createLogger } from '../utils/logger.js';
 import { attachmentManager, downloadFile } from '../file-transfer/inbound/index.js';
 import { messageLogger } from '../feishu/message-logger.js';
-import { FeishuFileHandler } from './platforms/feishu/feishu-file-handler.js';
-import { FeishuMessageSender } from './platforms/feishu/feishu-message-sender.js';
+import { FeishuFileHandler } from '../platforms/feishu/feishu-file-handler.js';
+import { FeishuMessageSender } from '../platforms/feishu/feishu-message-sender.js';
 import { TaskFlowOrchestrator } from '../feishu/task-flow-orchestrator.js';
 import { TaskTracker } from '../utils/task-tracker.js';
 import { BaseChannel } from './base-channel.js';
